backend/requirements.txt
Django>=1.8.6
django-cors-headers>=0.13
djangorestframework>=3.3.1
django-oauth-toolkit>=0.9.0
django-oauth2-provider>=0.2.3
djangorestframework-oauth>=1.0.1
gunicorn>=0.17.4
html2text>=2016.1.8
pytz>=2012f
